% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanboltr-package.R
\docType{data}
\name{baseApiUrl}
\alias{baseApiUrl}
\title{Base Oceanbolt API URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
baseApiUrl
}
\description{
Base Oceanbolt API URL
}
\keyword{internal}
