% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanboltr-package.R
\name{.onLoad}
\alias{.onLoad}
\title{Internal function to initialize package options}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{a character string giving the library directory where
the package defining the namespace was found}

\item{pkgname}{a character string giving the name of the package}
}
\value{
Invisible NULL
}
\description{
Internal function to initialize package options
}
\keyword{internal}
