% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCongestionTimeseries.R
\name{getCongestionTimeseries}
\alias{getCongestionTimeseries}
\title{Retrieves live congestion data for vessels currently congestion.}
\usage{
getCongestionTimeseries(
  regionId = c("BALTIC"),
  countryCode = c(),
  commodity = c("iron_ore_fines", "iron_ore_pellets", "iron_ore_unclassified",
    "magnetite"),
  commodityGroup = c(),
  segment = c("shortsea", "handysize", "supramax", "panamax", "capesize"),
  subSegment = c(),
  lastLoadCountry = c(),
  operation = c("discharge", "load", "yard", "unkown"),
  token = Sys.getenv("OCEANBOLT_TOKEN")
)
}
\arguments{
\item{regionId}{The list of regionIds to get congestion data for. Allowed
values can be obtained with \code{\link{listRegions}()}.}

\item{countryCode}{The list of countries to get congestion data for. Allowed
countries' codes can be obtained with \code{\link{listCountries}()}.}

\item{commodity}{The list of commodities to get congestion data for. Allowed
values can be obtained with \code{\link{listCommodities}()}).}

\item{commodityGroup}{The list of commodity groups to get congestion data
for. Allowed values can be obtained with \code{\link{listCommodities}()}).}

\item{segment}{The list of vessel segments for which to get fleet speed data.
Allowed values can be obtained with \code{\link{listSegments}()}.}

\item{subSegment}{The list of vessel sub segments for which to get fleet
speed data. Allowed values can be obtained with
\code{\link{listSegments}()}.}

\item{lastLoadCountry}{The list of load countries to get congestion data for.
Allowed countries' codes can be obtained with \code{\link{listCountries}()}.}

\item{operation}{The operational status of the vessel. By specifying this
filter it is possible to filter on the vessels that are waiting to discharge
('discharge'), the that are waiting to load ('load), the vessels that are
waiting to go to yard ('yard). Allowed values are
\code{"discharge", "load", "yard", "unkown"}.}

\item{token}{Oceanbolt API token, read from the environment by default.}
}
\value{
Data.table with data for vessels currently congestion.
}
\description{
Retrieves live congestion data for vessels currently congestion.
}
\details{
See \url{https://openapi.oceanbolt.com/#operation/getCongestionTimeseries}
for details.
}
\examples{
\dontrun{
congestionTimeseries <- getCongestionTimeseries()
congestionTimeseries
}

}
