% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPortCalls.R
\name{getPortCalls}
\alias{getPortCalls}
\title{Retrives port calls for either a list of imos or list of ports.}
\usage{
getPortCalls(
  imo = c(),
  portId = listPorts()[region == "BALTIC"]$portId,
  unlocode = c(),
  page = NULL,
  limit = 1000,
  token = Sys.getenv("OCEANBOLT_TOKEN")
)
}
\arguments{
\item{imo}{The list of unique vessel identifiers (IMO numbers). This allows
to filter tradeflows to show data only for a subset of vessels.}

\item{portId}{List of portIds to get data for. Allowed values can be obtained
with \code{\link{listPorts}()}.}

\item{unlocode}{List of UNLOCODES to get data for. Allowed values can be
obtained with \code{\link{listPorts}()}.}

\item{page}{The page number to retrieve of the response. If \code{NULL} all
pages are returned.}

\item{limit}{The number of voyages to retrieve per page (max is 1000,
default is 50) This parameter makes sense only if \code{page} in
no \code{NULL}.}

\item{token}{Oceanbolt API token, read from the environment by default.}
}
\value{
Data.table with trade flows.
}
\description{
Retrives port calls for either a list of imos or list of ports.
}
\details{
See \url{https://openapi.oceanbolt.com/#operation/vesselsPortCalls} for
details.
}
\examples{
\dontrun{
portCalls <- getPortCalls()
portCalls
}

}
