% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTonnageZoneCount.R
\name{getTonnageZoneCount}
\alias{getTonnageZoneCount}
\title{Retrieves tonnage data (vessel counts and dwt) for a given list of
zone_ids/segments/directions}
\usage{
getTonnageZoneCount(
  zoneId = 16,
  segment = c("shortsea", "handysize", "supramax", "panamax", "capesize"),
  subSegment = c(),
  direction = c("NNE", "ENE", "ESE", "SSE", "SSW", "WSW", "WNW", "NNW"),
  ladenStatus = c("laden", "ballast"),
  portStatus = c("in_port", "at_sea"),
  excludeMpv = TRUE,
  token = Sys.getenv("OCEANBOLT_TOKEN")
)
}
\arguments{
\item{zoneId}{The list of zoneIds to get tonnage data for. ZoneId can be
obtained either with \code{\link{listZones}()}, or alternatively from the
\url{https://app.oceanbolt.com} interactive zone selector. Defaults to
\code{"Baltic Sea"}.}

\item{segment}{The list of vessel segments for which to get fleet speed data.
Allowed values can be obtained with \code{\link{listSegments}()}.}

\item{subSegment}{The list of vessel sub segments for which to get fleet
speed data. Allowed values can be obtained with \code{\link{listSegments}()}.}

\item{direction}{The list of directions to get fleet speed data for.
The following directions are allowed: \code{NNE, ENE, ESE, SSE, SSW, WSW,
WNW, NNW}. Directions can also be obtained from the interactive direction
selector found at \url{https://app.oceanbolt.com}.}

\item{ladenStatus}{The laden status to get tonnage data for. The following
directions are allowed: \code{laden, ballast}.}

\item{portStatus}{The port status to get tonnage data for. The following
directions are allowed: \code{in_port, at_sea}.}

\item{excludeMpv}{Boolean, whether to exclude mpv vessels from the counts.}

\item{token}{Oceanbolt API token, read from the environment by default}
}
\value{
Data.table with average fleet speed for different
segments/zones/laden status.
}
\description{
Retrieves tonnage data (vessel counts and dwt) for a given list of
zone_ids/segments/directions
}
\details{
See \url{https://openapi.oceanbolt.com/#operation/getFleetSPeed} for details.

\emph{NOTE}: Parameter \code{format} is not used in this function and always set
to \code{"json"}.
}
\examples{
\dontrun{
speed <- getTonnageZoneCount()
speed
}

}
