% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTradeFlows.R
\name{getTradeFlows}
\alias{getTradeFlows}
\title{Retrieves all the individual voyages for the given filter parameters.
Response is paginated, and endpoint accepts a paging parameter to specify
which page to return. It is also possible to set the number of voyages to
return per query, default is 50 and maximum is 2000.}
\usage{
getTradeFlows(
  commodity = c("iron_ore_fines", "iron_ore_pellets", "iron_ore_unclassified",
    "magnetite"),
  direction = c("export", "import"),
  imo = c(),
  segment = c("shortsea", "handysize", "supramax", "panamax", "capesize"),
  subSegment = c(),
  fromDate = as.Date("2020-01-01"),
  toDate = Sys.Date(),
  fromCountryCode = c(),
  toCountryCode = c(),
  fromRegion = c(),
  toRegion = c(),
  excludeIntraCountry = TRUE,
  excludeUnknownDestinations = FALSE,
  page = NULL,
  limit = 1000,
  token = Sys.getenv("OCEANBOLT_TOKEN")
)
}
\arguments{
\item{commodity}{The commodity to get tradeflows for (get a list of all
commodities with \code{\link{listCommodities}()}).}

\item{direction}{The flow direction of cargo. This parameter only applies for
top ports/top country queries. Allowed values are
\code{c("export", "import")}.}

\item{imo}{The list of unique vessel identifiers (IMO numbers). This allows
to filter tradeflows to show data only for a subset of vessels.}

\item{segment}{The list of vessel segments for which to get fleet speed data.
Allowed values can be obtained with \code{\link{listSegments}()}.}

\item{subSegment}{The list of vessel sub segments for which to get fleet
speed data. Allowed values can be obtained with
\code{\link{listSegments}()}.}

\item{fromDate}{The starting date to get tradeflows from.}

\item{toDate}{The ending date to get tradeflows to.}

\item{fromCountryCode}{The list of two letter ISO country codes for loading
(export) countries. Full list of countries can be get with
\code{\link{listCountries}()}.}

\item{toCountryCode}{The list of two letter ISO country codes for discharge
(import) counties. Full list of countries can be get with
\code{\link{listCountries}()}.}

\item{fromRegion}{The list of regions for loading (export) countries.}

\item{toRegion}{The list of of regions for discharge (import) counties.}

\item{excludeIntraCountry}{Determines whether or not to exclude voyages that
start and end in the same country (default to \code{TRUE}).}

\item{excludeUnknownDestinations}{Determines whether or not to exclude
voyages that have unknown destinations (default to \code{FALSE}).}

\item{page}{The page number to retrieve of the response. If \code{NULL} all
pages are returned.}

\item{limit}{The number of voyages to retrieve per page (max is 2000, default
is 50). This parameter makes sense only if \code{page} in no \code{NULL}.}

\item{token}{Oceanbolt API token, read from the environment by default.}
}
\value{
Data.table with trade flows.
}
\description{
Retrieves all the individual voyages for the given filter parameters.
Response is paginated, and endpoint accepts a paging parameter to specify
which page to return. It is also possible to set the number of voyages to
return per query, default is 50 and maximum is 2000.
}
\details{
See \url{https://openapi.oceanbolt.com/#operation/postTradeflowLadenLegs} for
details.
}
\examples{
\dontrun{
tradeFlows <- getTradeFlows()
tradeFlows
}

}
