% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVesselStatus.R
\name{getVesselStatus}
\alias{getVesselStatus}
\title{Retrives live vessel status for any vessel or group of vessels in the dry
bulk fleet.}
\usage{
getVesselStatus(
  direction = c("NNE", "ENE", "ESE", "SSE", "SSW", "WSW", "WNW", "NNW"),
  ladenStatus = c("laden", "ballast"),
  imo = c(),
  loadCountry = c(),
  destinationCountry = c(),
  loadRegion = c(),
  destinationRegion = c(),
  zoneId = 16,
  segment = c("shortsea", "handysize", "supramax", "panamax", "capesize"),
  subSegment = c(),
  excludeUnknownDestinations = TRUE,
  excludeMpv = FALSE,
  token = Sys.getenv("OCEANBOLT_TOKEN")
)
}
\arguments{
\item{direction}{The list of directions to get fleet speed data for.
The following directions are allowed: \code{NNE, ENE, ESE, SSE, SSW, WSW,
WNW, NNW}. Directions can also be obtained from the interactive direction
selector found at \url{https://app.oceanbolt.com}.}

\item{ladenStatus}{The laden status to get tonnage data for. The following
directions are allowed: \code{laden, ballast}.}

\item{imo}{The list of unique vessel identifiers (IMO numbers). This allows
to filter vessels data to show data only for a subset of vessels.}

\item{loadCountry}{Sets load country filter for vessels. Full list of
countries can be obtained with \code{\link{listCountries}()}.}

\item{destinationCountry}{Sets destination country filter for vessels. Full
list of countries can be obtained with \code{\link{listCountries}()}.}

\item{loadRegion}{Sets load region filter for vessels. Full list of regions
can be obtained with \code{\link{listRegions}()}.}

\item{destinationRegion}{Sets destination region filter for vessels. Full
list of regions can be obtained with \code{\link{listRegions}()}.}

\item{zoneId}{The list of zoneIds to get tonnage data for. ZoneId can be
obtained either with \code{\link{listZones}()}, or alternatively from the
\url{https://app.oceanbolt.com} interactive zone selector. Defaults to
\code{"Baltic Sea"}.}

\item{segment}{The list of vessel segments for which to get fleet speed data.
Allowed values can be obtained with \code{\link{listSegments}()}.}

\item{subSegment}{The list of vessel sub segments for which to get fleet
speed data. Allowed values can be obtained with \code{\link{listSegments}()}.}

\item{excludeUnknownDestinations}{Boolean, whether to exclude vessels with
unknown destinations.}

\item{excludeMpv}{Boolean, whether to exclude mpv vessels from the counts.}

\item{token}{Oceanbolt API token, read from the environment by default.}
}
\value{
Data.table with vessels.
}
\description{
Retrives live vessel status for any vessel or group of vessels in the dry
bulk fleet.
}
\details{
See \url{https://openapi.oceanbolt.com/#operation/vesselsLiveStatus} for
details.
}
\examples{
\dontrun{
vesselStatus <- getVesselStatus()
vesselStatus
}

}
