% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCommodities.R
\name{listCommodities}
\alias{listCommodities}
\title{Retrieves a list of all commodities available in the trade flow database}
\usage{
listCommodities()
}
\value{
Data.table with available commodities
}
\description{
Retrieves a list of all commodities available in the trade flow database
}
\details{
See \url{https://openapi.oceanbolt.com/#operation/listCommodities} for
details.
}
\examples{
\dontrun{
commodities <- listCommodities()
commodities
}

}
