% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCountries.R
\name{listCountries}
\alias{listCountries}
\title{Retrieves all the countries that are available in the tradeflow database}
\usage{
listCountries()
}
\value{
Data.table with available countries
}
\description{
Retrieves all the countries that are available in the tradeflow database
}
\details{
See \url{https://openapi.oceanbolt.com/#operation/listCountries} for details.
}
\examples{
\dontrun{
countries <- listCountries()
countries
}

}
