% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPorts.R
\name{listPorts}
\alias{listPorts}
\title{Retrieves all the ports that are available in the congestion database}
\usage{
listPorts(...)
}
\arguments{
\item{...}{Extra parameters passed to \code{\link{listEntities}()}}
}
\value{
Data.table with available countries
}
\description{
Retrieves all the ports that are available in the congestion database
}
\details{
See \url{https://openapi.oceanbolt.com/#operation/listCongestionPorts} for
details.
}
\examples{
\dontrun{
ports <- listPorts()
ports
}

}
