% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listRegions.R
\name{listRegions}
\alias{listRegions}
\title{Retrieves all the regions that are available in the tradeflow database}
\usage{
listRegions(...)
}
\arguments{
\item{...}{Extra parameters passed to \code{\link{listEntities}()}}
}
\value{
Data.table with available regions
}
\description{
Retrieves all the regions that are available in the tradeflow database
}
\details{
See \url{https://openapi.oceanbolt.com/#operation/listRegions} for details.
}
\examples{
\dontrun{
regions <- listRegions()
regions
}

}
