% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSegments.R
\name{listSegments}
\alias{listSegments}
\title{Retrives list of available segments in database}
\usage{
listSegments(...)
}
\arguments{
\item{...}{Extra parameters passed to \code{\link{listEntities}()}}
}
\value{
Data.table with available segments
}
\description{
Retrives list of available segments in database
}
\details{
See \url{https://openapi.oceanbolt.com/#operation/listSegments} for details.
}
\examples{
\dontrun{
segments <- listSegments()
segments
}

}
