% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listZones.R
\name{listZones}
\alias{listZones}
\title{Retrieves all available tonnage zones}
\usage{
listZones()
}
\value{
Data.table with available tonnage regions
}
\description{
Retrieves all available tonnage zones
}
\details{
See \url{https://openapi.oceanbolt.com/#operation/listTonnageZones} for
details.
}
\examples{
\dontrun{
zones <- listZones()
zones
}

}
