% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanboltr-package.R
\docType{package}
\name{oceanboltr-package}
\alias{oceanboltr}
\alias{oceanboltr-package}
\title{oceanboltr: Wrapper for Oceanbolt API (https://openapi.oceanbolt.com/)}
\description{
Provides full access to Oceanbolt (https://openapi.oceanbolt.com/) API.
}
\details{
Wrapper for Oceanbolt API (https://openapi.oceanbolt.com/)
}
\author{
\strong{Maintainer}: Alexander Gedranovich \email{gedranovich@gmail.com} [contractor]

Other contributors:
\itemize{
  \item Mads Schou-Andreasen \email{info@oceanbolt.com} [copyright holder, funder, reviewer]
}

}
\keyword{internal}
