% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerToken.R
\name{registerToken}
\alias{registerToken}
\title{Registers Oceanbolt API token for further use}
\usage{
registerToken(token, type = "plain")
}
\arguments{
\item{token}{Raw value of the Oceanbolt API token}

\item{type}{One of the two acceptable types of token registration}
}
\value{
Invisible NULL unless stops with error
}
\description{
There are two types of access token registration: \code{plain} and
\code{keyring}. With \code{plain} you just pass token as-is and it is used
during your R session. \code{keyring} allows you to use your OS's keyring
backend (if enabled) which securely stores your token persistantly for all
further sessions.
}
\examples{
\dontrun{
# Register token as a plain text for the current session
registerToken("<YOUR_OCEANBOLT_TOKEN>")

# Register token with your OS's keyring for further usage
registerToken("<YOUR_OCEANBOLT_TOKEN>", type = "keyring")
}

}
