% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeApiCall.R
\name{safeApiCall}
\alias{safeApiCall}
\title{Internal function to make safe API calls with retries on rate limits}
\usage{
safeApiCall(httrFunction, ...)
}
\arguments{
\item{httrFunction}{Either \code{GET} or \code{POST} functions from
\code{httr} package}

\item{...}{Extra arguments for \code{httrFunction}}
}
\value{
Response object from \code{httrFunction}
}
\description{
Internal function to make safe API calls with retries on rate limits
}
\keyword{internal}
