% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listEntities.R
\name{listEntities}
\alias{listEntities}
\title{Retrieves a list of all entities of a specific type available
in the trade flow database}
\usage{
listEntities(
  entity = c("countries", "regions", "commodities", "zones", "segments")
)
}
\arguments{
\item{entity}{Type of the entity, one of \code{countries}, \code{regions},
\code{commodities}, \code{zones}, \code{segments}}
}
\value{
Data.table with available entities
}
\description{
This is a generic internal function to be used inside the package
}
\keyword{internal}
